* RTL toasts

+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/Notification Toast/User Input/Left to Right/Normal/VisualToastNotificationNumericUpDownUserInputForm.cs
@@ -266,7 +266,7 @@ namespace Krypton.Toolkit
             return base.ShowDialog(owner);
         }
 
-        internal static int ShowNotification(KryptonUserInputToastNotificationData data)
+        internal static decimal ShowNotification(KryptonUserInputToastNotificationData data)
         {
             var owner = data.ToastHost ?? FromHandle(PI.GetActiveWindow());
